Initial Changes From Base Repos
Re-worked all of the hardcoded directories and settings into constant files where possible so that the changes can be done in a single place and it can then be used by any of them.

Added a blank image as an example for users. Added a code section to cater for cases where Blank images are used so that the Blank properties are not added to the metadata of the NFT.

Updated the updateMetas.js file to no longer require the user to manually update the count of files that should be excluded from the update.

Fixed the preview function to work with custom_fields as the library will be used for multiple file uploads.

Removed the update_info.js file and replaced it with update_image_info and update_metadata_info so that if you upload your files separate from the NFTPort uploads, then you can use these files before minting your NFTs.

Testing still to be conducted for all files and functions along with uploading via NFTPort vs manually uploading to IPFS and using the update_image_info and update_metadata_info functions.

Rate limit has been added via a new library "async-sema" that allows for throttling per second.

Improvements were done on the Base repository and these have been re-worked into this code base. Minor changes were made to these, but the overall code base works the same.

The _ipfsMeta file is no longer used and instead the _metadata.json file is used throughout. This allows for both external file uploading processes as well as NFTPort file uploading processes to work together.

New script processes were added to allow for updating of specific metadata for the NFTs.

A new checking process has been added that will allow a user to run a check through all minted files and generate a list of failed mints where status is not OK or where the status is OK, the transaction can't be found online.

A few other minor changes were made as well like README updates and other minor enhancements to be more in line with the base code.

Made minor enhancements and fixes while testing each of the processes to ensure this can be used whether using NFTPort for files upload or an alternative like Pinata.

Ensured that both GenericJSON and Normal JSON, depending on Reveal requirements, will upload and mint without issues.