NFTPort API Scripts Retry Addition
Added retry functionality on the NFTPort API scripts.

A new value can be set under ACCOUNT_DETAILS which will be used as a retry limit when making API calls and the endpoint can't be reached or an unsuccessful response is returned.

Moved all variables in account_details.js file to be strings instead of numbers so that if they aren't set, the scripts won't break when they are running.