Json Migration And Error Enhancement
Added a new script that can be used to migrate the metadata_uri data from the _ipfsMetas.json file to the metadata.json file. Added a backup process for the files so that users can fall back to the original files.

Updated error checking sections within the check_mints.js process and the minting scripts. The re-mint.js script was not modified as it is assumed the check_mints.js was run before hand.

Added comments in files that got updated.