Addition of error handling post API call
Added error checking after minting API calls to notify user that there might be problems with the mint.

Added comments to the minting script files for other developers to follow

Moved the TIMEOUT setting that gets used when the API calls receive errors to the account_details.js file so that it can be set in a single place and used throughout all files. It can be overwritten within the file itself if really needed by changing the value at the top of the script file.