Rate Limit Default And Small Changes
Updated the value of 'x' to '1' for the max_rate_limit variable in the account_details.js file. This is to ensure that API calls goes through as a numeric value will be passed into the Rate limit function instead of 'NaN'.

Updated comments throughout various scripts.

Removed '[]' brackets in the nft_details.js file as it confuses people a bit.