Migration and normalization
Migrated scripts into sub-folders.

Added comments in files.

Reverted to using ipfs folder and file instead of updating the metadata.json file to keep in sync with the parent code bases.

Added new uploadGenericMeta variable to the account_details.js file that will allow a user to dictate if they want to upload the genericJSON directory files or the json directory files, depending on true or false.