Simple-Get And Canvas Patch
Security upgrade for simple-get library from version 3.0.3 to 4.0.1

Upgrade for canvas library from 2.8.0 to 2.9.0. Also added the security upgrade of simple-get library from 3.0.3 to 4.0.1 for this library.

Removed the requirement of running multiple npm install commands. The npm install command will install all the required dependencies and correct versions in a single command.