V1.6.2 Changes 1
Abstracted the FetchWithRetry function into a module so that it becomes re-useable.

Abstracted the GetMintAddress function into a module so that it becomes re-useable.

Re-worked all files to make use of the new abstracted modules.

Added a new uploadMetas_directory.js script to cater for the new NFTPort API call.

Modified README file with the new commands and added the package-lock.json and yarn.lock files to the .gitignore files.