V1.6.2 Changes 2 - Filtration And Dependency Rules
Renamed exclusions functionality to Filtration as this will be a broader naming convention for both inclusion and exclusion rules.

Added a new optional dependentTraits object to the layer configuration array.
If the this object is left out, then the dependent_traits script will be skipped.

Added a new dependent_traits script that will perform checks to ensure if layer item A is chosen, then layer items X, Y, Z, etc. is also in the same edition, otherwise reject the edition.

Console error logging has been updated for previous exclusion rules to note that filtration rules blocked the edition.