Selected Traits List Object Creation Moved
Moved the selectedTraitsList set creation to the beginning of the loop for each layer configuration object instead of as a global set for all configuration items. This allows users to filtrations specifically for each layer configuration object.

Previous behaviour allowed the selected traits from one layer configuration object to affect the the maximum repeatability configuration of another layer configuration, which should not be the case as each of these settings are set at a layer configuration level, not at a global level for all layer configuration objects.

Should users require the previous functionality, then a new issue can be created where a minor rework can be done on the maximum repeatability filter along with re-working the main module.