Collection Start Edition Number
Created a new startCollectionEditionFromNumber setting in the constants/nft_details.js file. This will allow users to specify the edition number that their collection will start creating from.

This is needed for users that want to make use of Collection Contracts on NFTPort as well as anyone that wants to start their collections from a different number.