Minor Bug Fix On Starting startCollectionEditionFrom Setting
A minor bug was identified when setting the startCollectionEditionFrom value to '0'.

The check that was done previously would check if startCollectionEditionFrom exists as a number, which technically is correct, however, the number '0', gets picked up as not existing. An additional check was added to cater for '0' and now all cases are generating collection numbers correctly now.