Addition Of Layer Global Maximum Repeatability Setting
Added a new "layerItemsMaxRepeatedTrait" setting that can be used to set the maximum repeatability for all items within the layer. If this is set to 2 for example and your layer has 8 layer items, then a maximum of 16 images will be generated. This allows users to set the maximum repeatability once off for a layer's items instead of specifying each layer item with the same maximum repeatability.

Updated rarity on example layer item names.

Updated the console logging on the uploadMetas_directory.js script so that the user is not bombarded by messages in the console.